requirements.txt
opencv-python~=4.10.0.84
opencv-contrib-python
ffmpeg-python~=0.2.0
ultralytics~=8.3.38
lapx
torchreid
gdown
tensorboard
hub-sdk
supervision~=0.25.0
torch~=2.5.1
numpy~=1.26.4
pillow~=11.0.0
transformers~=4.46.3
torchvision~=0.20.1
albumentations~=1.4.22

scipy~=1.14.1
click~=8.1.7
roboflow~=1.1.49
PyYAML~=6.0.2
matplotlib~=3.9.2
tqdm~=4.67.1
sahi~=0.11.20
